undepricated convert_mlt
Moved convert_mlt from depricated to non-deprecated.
